% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{Account}
\alias{Account}
\title{ScrumSaga API Wrapper}
\usage{
Account(email, password)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Account()
}
\keyword{api}
