% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagaAccount.R
\name{github_api}
\alias{github_api}
\title{ScrumSaga API Wrapper}
\usage{
github_api(path)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
github_api()
}
\keyword{api}
