% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{Project}
\alias{Project}
\title{ScrumSaga API Wrapper}
\usage{
Project(account, repo_namespace, repo_email, repo_name)
}
\description{
Container and associated functionaity for Projects
}
\examples{
Project()
}
\keyword{api}
